requirements.txt
sqlalchemy
psycopg2-binary
alembic
python-dotenv
fastapi
uvicorn
pydantic
